+++ cpp/src/scheduler/Utils.cpp
@@ -4,16 +4,17 @@
  * Proprietary and confidential.
  ******************************************************************************/
 
-#include <chrono>
 #include "Utils.h"
+
+#include <chrono>
+
 
 namespace zilliz {
 namespace milvus {
 namespace engine {
 
 uint64_t
-get_current_timestamp()
-{
+get_current_timestamp() {
     std::chrono::time_point<std::chrono::system_clock> now = std::chrono::system_clock::now();
     auto duration = now.time_since_epoch();
     auto millis = std::chrono::duration_cast<std::chrono::milliseconds>(duration).count();